add cookie function on main page
