--- public/front_js/cookie.js
@@ -1,32 +1,0 @@
-export let Cookie={
-    setCookie : function(name, value, days){
-        if (days) {
-                var date = new Date();
-                date.setTime(date.getTime() + (days * 24 * 60 * 60 * 1000));
-                var expires = "; expires=" + date.toGMTString();
-        } else {
-               var expires = "";
-        }
-        
-        document.cookie = name + "=" + value + expires + "; path=/";
-    },
-    getCookie : function(name) {
-        var i, x, y, ARRcookies = document.cookie.split(";");
-        
-        for (i = 0; i < ARRcookies.length; i++) {     
-                x = ARRcookies[i].substr(0, ARRcookies[i].indexOf("="));
-                y = ARRcookies[i].substr(ARRcookies[i].indexOf("=") + 1);
-                
-                x = x.replace(/^\s+|\s+$/g, "");
-    
-                if (x == name) {
-                        return unescape(y);
-                }
-        }
-    }
- 
-};
-
-
-
-
