+++ views/1_trabel.html
@@ -7,8 +7,9 @@
     <link rel="stylesheet" href="/css/1_trabel.css">
     <link rel="icon" type="image/png" href="image/favicon.png"/> 
     <!-- href에 이름만 바꾸면 됨 -->
+    <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
     <script src="front_js/colors.js"></script>
-    <script src="front_js/visible_control.js"></script>
+    <script src="front_js/mainPage.js"></script>
 </head>
 
 <body>
@@ -27,7 +28,7 @@
     </div>
 
     <div class="input">
-        <form action="/place" method="get" onKeyDown="if(event.keyCode == 13) {return false;}"">
+        <form action="" method="get" onKeyDown="if(event.keyCode == 13) {return false;}"">
             <p>
                 <input type="text" name="inputPlace" class="input_place" placeholder="여행 지역을 입력해주세요." onKeyDown= "placeBoxEvent()">
             </p>
@@ -45,7 +46,7 @@
             <p>
                 설정이 완료되었으면 아래 버튼을 눌러주세요.
                 <br>
-                <input type="submit" value="확인">
+                <input onclick="Cookie();" type="submit" value="확인">
             </p>
         </form>
     </div>
